выполнить ТЗ спринта 8

+++ src/main/java/ru/yandex/practicum/filmorate/model/Film.java
@@ -11,5 +11,5 @@ public class Film {
     private final String name;
     private final String description;
     private final LocalDate releaseDate;
-    private final int duration;
+    private final Long duration;
 }
